NN sample prebuilt: x86_64|arm64: Use page agnostic target
The prebuilts for arm64 and x86_64 need to be have ELF p_align at
least as large as 16kB in order to support both 4kB and 16kB page-size
targets.

To achieve this we use the page agnostic targets to generate prebuilts
for arm64 and x86_64.

While at it also update the default target for other architectures to
be compliant with trunk stable format:

   Must be of the form <product>-<release>-<variant>

Bug: 312573097
Test: Boot 16k page size device.
Change-Id: I9a87ac506f227753471536427ca5e4a834b4439e
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>